+++ src/main/java/net/raphimc/audiomixer/soundmodifier/impl/SpatialModifier.java
@@ -41,7 +41,7 @@ public class SpatialModifier implements SoundModifier {
     }
 
     @Override
-    public void modify(final AudioFormat audioFormat, final int[] renderedSamples) {
+    public void modify(final AudioFormat audioFormat, final float[] renderedSamples) {
         if (audioFormat.getChannels() != 2) {
             throw new UnsupportedOperationException("Target audio format must have 2 channels");
         }
@@ -52,8 +52,8 @@ public class SpatialModifier implements SoundModifier {
         }
 
         for (int i = 0; i < renderedSamples.length; i += 2) {
-            renderedSamples[i] = (int) (renderedSamples[i] * (1F - this.panning) * this.attenuation);
-            renderedSamples[i + 1] = (int) (renderedSamples[i + 1] * this.panning * this.attenuation);
+            renderedSamples[i] = renderedSamples[i] * (1F - this.panning) * this.attenuation;
+            renderedSamples[i + 1] = renderedSamples[i + 1] * this.panning * this.attenuation;
         }
     }
